refactor: move schema files into directory
