+++ drizzle.config.ts
@@ -2,9 +2,9 @@ import { defineConfig } from 'drizzle-kit';
 
 export default defineConfig({
   dialect: 'postgresql',
-  schema: './src/db/schema.ts',
+  schema: './src/db/schema',
   out: './supabase/migrations',
-  schemaFilter: ['public'],
+  schemaFilter: ['public', 'risks'],
   introspect: {
     casing: 'camel'
   },
