+++ src/db/schema/risks.ts
@@ -59,8 +59,8 @@ export const risks = risksSchema.table('risk_entries', {
   source: text(),
   image: text('image_url'),
   articleDate: timestamp('article_date'),
-  category: riskCategoryEnum(),
-  finanicalImpat: bigint('financial_impact', { mode: 'number' }),
+  riskCategory: riskCategoryEnum('risk_category'),
+  finanicalImpact: bigint('financial_impact', { mode: 'number' }),
   summary: text(),
   mitigation: text(),
   justification: text(),
