+++ src/lib/server/shared/auth-check.ts
@@ -3,7 +3,7 @@ import 'server-only';
 import { type User } from '@supabase/supabase-js';
 import { and, eq, inArray } from 'drizzle-orm';
 import { db } from '#/db';
-import { roles } from '#/db/schema';
+import { roles } from '#/db/schema/public';
 import { RoleEnum } from '#/enums';
 import { createClient } from '#/lib/supabase/server';
