+++ src/lib/types.ts
@@ -1,5 +1,5 @@
-import type { users } from '#/db/auth-schema';
-import type { roles } from '#/db/schema';
+import type { users } from '#/db/schema/auth';
+import type { roles } from '#/db/schema/public';
 
 export type Member = Omit<typeof users.$inferSelect, 'userMetadata'> &
   Pick<typeof roles.$inferSelect, 'role' | 'active'> &
