+++ src/trpc/routes/org/org-invite-create.ts
@@ -3,8 +3,8 @@ import 'server-only';
 import { TRPCError } from '@trpc/server';
 import { and, eq } from 'drizzle-orm';
 import { db } from '#/db';
-import { users } from '#/db/auth-schema';
-import { invites, roles } from '#/db/schema';
+import { users } from '#/db/schema/auth';
+import { invites, roles } from '#/db/schema/public';
 import { RoleEnum } from '#/enums';
 import authCheck from '#/lib/server/shared/auth-check';
 import { authProcedure } from '#/trpc/init';
